% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRC}
\alias{DRC}
\title{DRC Seaosnal data to match regions from Parr et al. 2016 JID}
\format{
An object of class \code{data.frame} with 8 rows and 26 columns.
}
\usage{
data(DRC)
}
\description{
DRC Seaosnal data to match regions from Parr et al. 2016 JID
}
\keyword{datasets}
