% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq.0}
\alias{eq.0}
\title{Equilibrium conditions for model. Also stored internally for model use when
package is not loaded}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(eq.0)
}
\description{
Equilibrium conditions for model. Also stored internally for model use when
package is not loaded
}
\keyword{datasets}
