% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrp2-simulation.R
\name{hrp2_Simulation}
\alias{hrp2_Simulation}
\title{Individual Based Malaria Simulator}
\usage{
hrp2_Simulation(ID = NULL, root = NULL, direct.input = NULL,
  years = 0.5, time.step = 1, states = c(1, 2, 3, 4, 5, 6),
  storage = 50, storage_capture = 30, just_storage_results = FALSE,
  N = 1000, max.age = 60, prev.0 = 0.3, strains.0 = 9,
  desired.freq = NULL, d.death = 21, EIR = 200/365, a.0 = 8 * 365,
  rho = 0.85, zeta.sd = 1.67, omega = 0.8405912, theta = 1, ft = 0.4,
  rdt.det = 1, rdt.nonadherence = 0, microscopy.use = 0,
  include.nmf = FALSE, fever.age.breaks = c(-0.1, 1, 2, 3, 4, 5, 7, 9, 11,
  13, 15, 20, 25, 30, 101), annual.nmf.per.age.bracket = c(2.235, 1.841,
  1.503, 1.175, 0.928, 0.67, 0.486, 0.512, 0.475, 0.429, 0.652, 0.747, 0.804,
  0.789), nmf.multiplier = 1, ss = NULL, fitness = 1, mu.0 = 0.132,
  beta = c(0.05879106, 0.07108894), a.k = 0.30677, delay.mos = 10,
  delay.gam = 12, d.E = 12, d.I = 200, d.T = 5, d.D = 5, d.U = 110,
  d.P = 25, d.A = d.I - d.D, d.1 = 0.160527, d.ID = 10 * 365,
  I.D0 = 1.577533, k.D = 0.476614, u.D = 9.44512, a.D = 8001.99,
  f.D0 = 0.007055, g.D = 4.8183, b.0 = 0.590076, b.1 = 0.5, d.B = 10 *
  365, I.B0 = 43.8787, k.B = 2.15506, u.B = 7.19919, phi.0 = 0.791666,
  phi.1 = 0.000737, d.CA = 10950, I.C0 = 18.02366, k.C = 2.36949,
  u.C = 6.06349, P.M = 0.774368, d.CM = 67.6952, g.1 = 1.82425,
  c.S = 0, c.D = 0.0676909, c.T = 0.322 * c.D, c.U = 0.0062, c.A = 0,
  c.P = 0)
}
\arguments{
\item{years}{Length of Simulation (years). Default = 0.5}

\item{time.step}{Simulation time interval considered in days. Has to be 1.}

\item{states}{Vector of the compartmental states. Default = c(1, 2, 3, 4, 5, 6), which are c("S","D","A","U","T","P")}

\item{storage}{Number of last time steps to store. If storage = 0 (default), then all results are stored}

\item{storage_capture}{Number of previous time steps used to calculate mean measures for series. Default = 30}

\item{just_storage_results}{Boolean for keeping just the storage results.
## Demographic paramaters}

\item{N}{Population Size. Default = 1000}

\item{max.age}{Maximum age (years). Default = 100}

\item{prev.0}{Assumed starting prevalence. Default = 0.3}

\item{strains.0}{Maximum starting MOI possible. Default = 5}

\item{desired.freq}{Desired hrp2 deletion frequency to be enforced when a previous simulation is reloaded}

\item{d.death}{Average life expectancy (years). Deafult = 21
## Epidemiological Parameters}

\item{EIR}{Maximum EIR / day. Default = 200/365,}

\item{a.0}{Age-dependent biting parameter (days). Default = 8*365}

\item{rho}{Age-dependent biting parameter (days). Default = 0.85}

\item{zeta.sd}{sd of lognormal distribution used to generate indiviudal specific biting. Default = 1.67}

\item{omega}{BM age scaling parameter. Default = 0.75
## Entomological parameters}

\item{theta}{Seasonal effects. Default = 1}

\item{ft}{Probability of receiving assesment. Default = 0.4}

\item{rdt.det}{Probability of detecting malaria given mono infection of hrp2' strains. Default = 1}

\item{rdt.nonadherence}{Probability of non adherence to test result. Default = 0, i.e. perfect adherence}

\item{microscopy.use}{Probability of microscopy use. Default = 0, i.e. no microscopy}

\item{include.nmf}{Boolean detailing whether to include nmf section. Default = FALSE}

\item{fever.age.breaks}{Age breaks for fevers}

\item{annual.nmf.per.age.bracket}{Annual number of non malarial fevers for each age bracket in fever.age.breaks.
Default is mean across 5 representative surveys from Burundi 2012, Liberia 2009/2011 and Nigeria 2010/2015}

\item{nmf.multiplier}{Multiplication for annual.nmf.per.age.bracket to introduce sensitivity. Default = 1
## delays and durations}

\item{mu.0}{Baseline mosquito mortality. Default = 0.132}

\item{beta}{Birthrate in form of y = beta[1]*EIR + beta[2]. Default = c(0.0572289,0.0696121)}

\item{a.k}{Daily biting rate. Default = 0.30667
## Diagnostics and nmf}

\item{delay.mos}{Mosquito extrinsic incubation period (days). Default = 10}

\item{delay.gam}{Delay from emergence of blood-stage parasites to onward infectivity (days). Default = 12.5}

\item{d.E}{Latent period delay in pre-erythrocytic stage (days). Default = 12}

\item{d.I}{Patent infection duration with no immunity (including disease) (days). Default = 200}

\item{d.T}{Clinical disease duration with treatment (days). Default  = 5}

\item{d.D}{Clinical disease duration without treatment (days). Default  = 5}

\item{d.U}{Sub-patent infection duration (days). Default = 100}

\item{d.P}{Prophylaxis with SP following treatment duration (days). Default = 25}

\item{d.A}{Asymptomatic infection duration (days). Default = d.I - d.D
## Pre-erythrocytic immunity parameters}

\item{d.1}{= 0.161 Probability with maximum immunity}

\item{d.ID}{= 10*365 Inverse of decay rate}

\item{I.D0}{Scale parameter. Default = 1.58}

\item{k.D}{Shape parameter. Default = 0.477}

\item{u.D}{Duration in which immunity is not boosted (days). Default = 9.45}

\item{a.D}{Scale parameter relating age to immunity (days). Default = 21.923*365}

\item{f.D0}{Parameter relating age to immunity. Default = 0.0071}

\item{g.D}{Shape parameter relating age to immunity. Default = 4.81
## Blood stage immunity parameters}

\item{b.0}{Probability with no immunity. Default = 0.590}

\item{b.1}{Maximum relative reduction. Default = 0.5}

\item{d.B}{Inverse of decay rate (days). Default = 10*365}

\item{I.B0}{Scale parameter. Default = 43.9}

\item{k.B}{Shape parameter. Default = 2.16}

\item{u.B}{Duration in which immunity is not boosted (days). Default = 7.2
## Acquired and maternal immunity parameters}

\item{phi.0}{Probability with no immunity. Default = 0.7916}

\item{phi.1}{Maximum relative reduction. Default = 0.00074}

\item{d.CA}{Inverse of decay rate (days). Default = 30*365}

\item{I.C0}{Scale parameter. Default = 18.0}

\item{k.C}{Shape parameter. Default = 2.37}

\item{u.C}{Duration in which immunity is not boosted (days). Default = 6.06}

\item{P.M}{New-born immunity relative to mother’s. Default = 0.774}

\item{d.CM}{Inverse of decay rate of maternal immunity (days). Default = 67.7
## contributions to infectious reservoir by state and age}

\item{g.1}{Relates infectiousness probability of detection. Default = 1.82}

\item{c.S}{Contribution from S. Default = 0}

\item{c.D}{Contribution from D. Default = 0.068}

\item{c.T}{Contribution from T. Default = 0.32 * c.D}

\item{c.U}{Contribution from U. Default = 0.0062}

\item{c.A}{Contribution from A. Default = c.U + (c.D - c.U)*q^g.1, where q is
the probability of detection of parasites as calculated using pre-erythrcytic immunity and age}

\item{c.P}{Contribution from P. Default = 0}
}
\value{
Result a list of 59 elements including the following
Time - Vector of time steps considered
Status - Matrix of individuals' infection status by Time
Last_Infection_Time - Vector of individual's last infection
Last_Bite_Time - Vector of individual's last bite
Age - Vector of individual's ages
N.Sens - Matrix of individuals' MOI of RDT sensitive strains by Time
N.Dels - Matrix of individuals' MOI of RDT resistant strains by Time
I.B - Vector of individual's blood stage immunity
I.CA - Vector of individual's acquired immunity
I.CM - Vector of individual's maternal immunity
I.D - Vector of individual's pre-erythrocytic immunity
Counter - Simulation step counter
d.EIR - daily EIR in individuals > 18 years
I.Reservoir - Vector of each strains weighted contribution to the last step's I.Reservoir
Sv - Susceptible mosquito population
Ev - Exposed mosquito population
Iv - Infected mosquito population
Prev - Daily prevalence
N.Sens/N.Dels - sum of sensitive to diagnostic and deletion strains
Incidence - Clinical cases / population / per day
Mono - monoinfected individual proportions
FOIv - Force of infection to vectors
S. ... - Series storage of various of the above
}
\description{
\code{hrp2_Simulation} simulates an individual based malaria system considering
multiplicity of infection and infection status of individuals.
}
\details{
The simulator considers individuals to progress through the Griffin 2016 model.

The model also introduces a further acknowledgement of the hrp2 strains present in hosts, and
the effects of this on detection.

The model first allocates a fixed length for all storage and proceeds in a ring like fashion allowing for delays to
be incorporated.

## Simulation parameters
}
