% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hrp2malaRia_package}
\alias{hrp2malaRia_package}
\title{hrp2malaria: Individual multiplicity of infection malaria model}
\description{
hrp2malaria: Individual multiplicity of infection malaria model
}
\author{
OJ Watson \email{oj.watson@hotmail.co.uk}
}
