% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_examples.R
\name{run_seasonality_shiny}
\alias{run_seasonality_shiny}
\title{Seasonality Shiny}
\usage{
run_seasonality_shiny()
}
\description{
\code{run_seasonality_shiny} runs a shiny app that displays the outcomes 
  of the seasonality investigation. The shiny shows how the bias in the 
  observed prevalence of pfhrp2 deletions changes for each level 1 admin 
  region changes throughout the transmission season
}
